feat: add include path to openssl in vendored mode
If OpenSSL's headers are not installed in a standard path,
compiling yara in vendored mode with YARA_CRYPTO_LIB=openssl is
not possible, as there is no way to provide the include path for
OpenSSL (only OPENSSL_LIB_DIR is available, but no
OPENSSL_INCLUDE_DIR).

Fix this by:

- Handling OPENSSL_DIR, from which the include and lib dirs are
  extrapolated.
- Otherwise, handle OPENSSL_LIB_DIR (already done) and
  OPENSSL_INCLUDE_DIR (addition from this MR).

This is done so that the behavior aligns with the rust bindings
for openssl: https://docs.rs/openssl/latest/openssl/#manual

This way, if a crate uses both the openssl and the yara crate,
setting those env variables will work for both.